// src/types.ts
export interface ResponseData {
  data: {
    id: string;
    symbol: string;
    name: string;
    nameid: string;
    rank: number;
    price_usd: string;
    percent_change_24h: string;
    percent_change_1h: string;
    percent_change_7d: string;
    price_btc: string;
    market_cap_usd: string;
    volume24: number;
    volume24a: number;
    csupply: string;
    tsupply: string;
    msupply: string;
  }[];
  info: {
    coins_num: number;
    time: number;
  };
}

export type FetchStatusState =
  | { status: "idle" }
  | { status: "loading" }
  | { status: "success"; data: ResponseData }
  | { status: "failure"; error: string };